+++ app/controllers/application_controller.rb
@@ -4,7 +4,8 @@ class ApplicationController < ActionController::Base
   protect_from_forgery with: :exception
 
   rescue_from CanCan::AccessDenied do |exception|
-    redirect_to root_path, :alert => exception.message
+    flash[:alert] = "你沒有瀏覽此頁面的權限!"
+    redirect_to root_path
   end
 
   def after_sign_in_path_for(resource)