+++ app/models/ability.rb
@@ -2,10 +2,8 @@ class Ability
   include CanCan::Ability
 
   def initialize(user)
-    user ||= User.new # guest user (not logged in)
-    if user.has_role? :admin
-      can :manage, :all
-    end
+    user ||= User.new # guest user
+    can :manage, :all if user.has_role? :admin
     # Define abilities for the passed in user here. For example:
     #
     #   user ||= User.new # guest user (not logged in)