+++ config/routes.rb
@@ -15,5 +15,7 @@ Bigheart::Application.routes.draw do
     end
   end
 
-  root :to => "home#index"
+  post 'send_mail/report'
+
+  root :to => "stories#index"
 end