+++ db/migrate/20131221064705_add_contact_email_to_stories.rb
@@ -1,0 +1,5 @@
+class AddContactEmailToStories < ActiveRecord::Migration
+  def change
+    add_column :stories, :contact_email, :string
+  end
+end