+++ db/migrate/20131225113547_change_appear_day_type.rb
@@ -1,0 +1,5 @@
+class ChangeAppearDayType < ActiveRecord::Migration
+  def change
+    change_column :stories, :appear_day, :text
+  end
+end